+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Changes to functions
 
+* `find_parameters()` now also works for `BFBayesFactor` objects.
 * Suppress non-informative warning in `get_data()` for model data with weights.
 * `format_value()` automatically uses scientific notation for *very* large numbers (> 1e+5). Furthermore, the check for integer values was made more robust, to avoid warnings when checking *very* large numbers for integer type.